requirements.txt
Django==1.10.5
django-extensions==1.7.5
djangorestframework==3.5.3
six==1.10.0
wheel==0.24.0
